Added some more meat to the README
